[bt][a2dp] Start streams that are opened
In some cases, the peer will open a StreamEndpoint,
complete opening it, and then fail to start the stream.
They are not required to, so this is fine, but it means
the stream will sit indefinitely in an opened state.

After a timeout (currently 2 seconds) attempt to start the
stream instead if it was opened remotely.

Bug: 50465
Test: added in bt-a2dp-unittests
Change-Id: I2132a276a05c8001a80677fcf451e594a597cc43
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435076
Reviewed-by: Jeff Belgum <belgum@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Ani Ramakrishnan <aniramakri@google.com>