[shell] Add error alternatives
The Alt combinator will discard a parse result if its second result
succeeds. This might be undesirable in the case of things like
Alt(Expression, Empty) because the failed parse of Expression might
contain useful error information, and if the parse is destined to fail,
that may be the point where we want it to fail.

Alt will now store the failed first parse result in the returned result,
if it successfully parsed more text than the succeeding parse. The Seq
combinator can then use this alternative return to effectively backtrack
if an error correction is needed.

Change-Id: I61d22097b94790a7a69e2004736951a8e023f206
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/390514
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Testability-Review: Vincent Belliard <vbelliard@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>