[ui-input] input-synthesis: move LegacyInjector to legacy_backend module
Now that we've laid the groundwork, move LegacyInjector to the
legacy_backend module. This means that synthesizer.rs is now agnostic
to which protocol is used to register a fake input device.

Note that synthesizer.rs is not yet agnostic to the protocol used to
represent an input device (said differently: synthesizer.rs still
depends on `fuchsia.ui.input.InputDevice`). We'll address this in
future CLs.

Bug: 60789
Test: fx test input-synthesis-tests
Change-Id: Ie7c3529ae4c0df85be4e127dda6812dfab26af51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/435014
Commit-Queue: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>