[rust][fasync] Retry: Integrate instrumentation module into executor
[Retrying with timer-independent fix which removes non-determinism
from the test which previously flaked in some cases.]

Integrate the instrumentation module into the
fuchsia-async executor, so that stats are collected at runtime.

Add a method to the collector to generate a snapshot from
the current state of the collector.

Bug: 58578
Change-Id: I9352388912bacc2ad9b4e904733e409562f8ed23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438360
Reviewed-by: Tyler Mandry <tmandry@google.com>
Testability-Review: Didrik Nordstrom <dnordstrom@google.com>
Commit-Queue: Didrik Nordstrom <dnordstrom@google.com>