[escher] Fix UB triggered by calling memcpy with a nullptr
Calling memcpy with a null argument is undefined behavior, even if the
number of bytes you request to copy is 0.  This patch fixes escher to
avoid calling memcpy if the number of items to copy is zero.

Test: UBSan
Bug: 60251
Fixed: 62135
Change-Id: Ie1334a982839b4e6db93f8d172c45a4ab934474e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/439876
Reviewed-by: Josh Gargus <jjosh@google.com>
Testability-Review: Drew Fisher <zarvox@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>