[fint] Call gn gen from fint set
Without any args for now, to keep this CL small. I'll start adding args
in a follow-up.

The fuchsia-x64-release-fint builder is still not expected to pass with
this change, because we're not yet setting the necessary args to produce
all the artifacts that infra expects from gn.

Bug: 61977
Change-Id: Idacc4bc9c7c37a8413421cdca5bc7e42b5043a73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438962
Testability-Review: George Kulakowski <kulakowski@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>