[testrunner] Execute a SEV instruction during SSH diags
Trigger a kernel command which will cause a core to execute a SEV
instruction (send event) when running the diags attempting to figure
out why SSH is non-responsive.

Some of the recent diags results indicate that one of the causes of
SSH being non-responsive might be the issue we have with cores
apparently becoming stuck in WFE on Sherlock while attempting to
obtain a kernel spinlock.

If this is the case, then sending an event with SEV should unstick the
core, and we should immediately see its thread queue drain.

Bug: 59720
Change-Id: Ifff2843e8e9eb90043536233f8cb51ec28712f2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437974
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>