[e2e] Add a timezone test for flutter
This is similar to fxr/367445, but using the flutter runner this time.
While the mechanims for obtaining the locales are identical to those in
the dart test, the runner is still a separate binary so could be subject
to independent variation. So we redo the testing.

This commit factors the common parts of the test into a separate rust
library.

Bug: 43341
Change-Id: Icb8fb1c83bfb89dd8978a6d877dd3e7d2b5dfe36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391701
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Testability-Review: Filip Filmar <fmil@google.com>