[settings] Add broker filter in policy proxy and implement direct responses
This CL filters the messages sent to policy proxies for only messages
that are sent by the switchboard and that are requests for a particular
setting type. These messages are forwarded to the policy handler to take
action.

The pass through case is trivially implemented by doing nothing with the
message, so that it continues onto the setting handler as normal. The
direct response case is also implemented, where the policy handler
directly gives a response to the client and the setting handler never
receives a message. The modified request case where the policy handler
modifies the request and passes it onto the setting handler is not yet
implemented.

Bug: 59747
Test: fx test setui_service_tests
MULTIPLY: setui_service_tests
Change-Id: I7b6a5a4b0e38001220481dc4ba4eb4f49a745bfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438577
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Testability-Review: Bryce Lee <brycelee@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>