[fidlgen][cpp] Use FullDecl to disambiguate types
In C++, it is ambiguous to use the name `Foo` if there is both a method
and a class named `Foo` in scope. The existing approach in the codegen
to disambiguate them attempts to prefix the class ones with `class Foo`,
but won't work for all cases (e.g. it would treat `::std::unique_ptr<Foo>`
as already fully qualified when it isn't). An easier alternative is to
just always refer to types by their fully qualified name.

Change-Id: Ic9d1c92df1a8c484c9501d88a1b9b6420a7d34ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438383
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>