[build] Ensure fidl-format and fidl-lint are always in the build graph.
The devshell commands 'fx format-code' and 'fx lint' depend on the
host tools 'fidl-format' and 'fidl-lint' respectively, and will
invoke Ninja to build them if they are not available or out of date.

However, this can only work correctly if these tools are in the build's
dependency graph, and this wasn't the case before this CL, causing
weird error messages when the commands were invoked. For more details
see the associated bug entry.

This CL fixes the issue by ensuring that these tools are always
reachable from the //default target, and thus that "fx gen" will
always generate a Ninja target for them.

For testing, the CL fixes the following case:

  1) fx --dir out/qemu-x64 set core.qemu-x64 --release \
      --with sdk/lib/fdio:tests \
      --with src/connectivity/network:tests \
      --with third_party/go:tests \
      --with src/lib/component:tests

  2) fx set-build-dir out/qemu-x64

  3) Modify a .fidl file somewhere.

  4) Call 'fx format-code'.

Fixed: 60822
Testing: See above
Change-Id: I61f36490c1175862b23007a405bc6d0e359dc8fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/431302
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Turner <digit@google.com>