[perfmon][trace_converters] Don't replace LBR records with heap garbage
The calculation of the size of a LBR record uses + instead of * to
calculate an array size. Additionally, the chromium exporter is storing
LBR records in a vector, which can't work as they are variable size.

Also adds a few missing fields of each LBR record to the exported JSON.

Change-Id: I73d1c2ce097b3b8f0012c1ad3bdfbc645c7cdb77
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/438517
Commit-Queue: Theodore Dubois <tbodt@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>