[fidl][benchmarks] Setup benchmarks for fidlc
During the union to xunion migration, there were performance
regressions in FIDLC for certain inputs. This CL adds
benchmarks to help track and quantify the regressions.
The JSON generation path is the target.

In addition to tracking regressions, there are new usages
of fidlc such as language server support for IDEs that
depend on fast compilation performance and this opens
the door to optimizing for that case.

Only a single benchmark was added and chromeperf pathnames
were chosen under the assumption that only the JSON
generation path is being tracked (no information on which
path in the name). This tests compilation on fuchsia
which might have different performance than on host, but
it should be similar.

Initial value:
fuchsia.fidlc_microbenchmarks: Struct/16  585483 +/- 3367 ns

Change-Id: I232dbcc0eb7aaa53782880e4e7685bdaa084550b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440055
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>