[bt][rfcomm] Add support for establishing outgoing RFCOMM channels
Adds a new method to initiate channel establishment.

In general, when opening a remote channel, we must 1) Make sure the
session multiplexer is started, 2) Negotiate the parameters for the DLC
(and transitively the session) by sending a PN Mux Command and 3) Open
the channel by sending an SABM command over the user DLCI.

Once the remote peer positively acknowledges with a UA response, we can
open and establish the channel and relay it to the profile client.

Bug: 59585
Test: Adds a test that verifies the complete open remote channel path to
bt-rfcomm-tests

Change-Id: I48ffd4f2c6344e3cbd1545bc2f22eb589d1f0097
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/432495
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>