[fidl][hlcpp] Flexible bits support
This change is very similar to the one for LLCPP i.e.
8dfdc60b554992c130ed50e027366d39962846e7. I largely copied the bits
golang templates and unit tests from LLCPP. The main differences are:

- HLCPP strict bits need to stay as a `enum class` for now, because of
breaking changes otherwise. The missing APIs from strict bits were
removed from the unit tests.

- Removed `AllowingUnknown` in favor of `static_cast`, and we intend to
do the same for LLCPP.

- Renamed `MyBits::mask` to `MyBits::kMask` to follow C++ style guide,
and we intend to do the same for LLCPP.

In order to support the bits API under C++14, language version macros
are used to conditionally define the `constexpr` member variables in the
header when under C++17, and in the implementation file when under
C++14.

Fixed: 61395

Change-Id: I0face4d68cca37ba257e9cc09b61d962be3f243b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441379
Reviewed-by: Felix Zhu <fcz@google.com>
Testability-Review: Felix Zhu <fcz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>