[build] have action() use the in-tree python interpreter
The use of a custom template python_action could not be applied to
GN files that are shared between the fuchsia build and other builds
which don't have that template (e.g. third_party/protobuf/BUILD.gn).

This also is nice because users don't have to remember to use a special
template.

Bug: 60924
Change-Id: Ia0e08209fbebb26de4c43de54f7f899673fa8e8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441212
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Gary Miguel <garymm@google.com>