[fvm] Support reserving slices in fvm host tool
Reserved slices are held in a dummy partition with a particular GUID.
This partition can be arbitrarily reclaimed by FVM later for its own
purposes.

The dummy partition will be hidden by the FVM driver; that is, no block
device is bound for the dummy partition.

This is necessary to reserve slices for the upcoming A/B snapshot
effort, which requires some extra metadata.

Adds a new flag to the sparse format marking a partition as an internal
reservation partition, since the sparse format doesn't have GUIDs. The
flag is translated to the appropriate GUID in fvm_sparse_image_reader.
Note that fvm_sparse_image (which generates sparse images) is *not*
updated in this patch, which is okay because that code is only used by
tests for now.

This change does not include new unit tests for fvm-host, the change was
manually tested. The plan going forward for testability is to verify the
migration code in the FVM driver which detects this reserve partition
and takes over the slices for other uses, once that migration code is
implemented.

Bug: 62062
Tested: Booted on a variety of devices.
Change-Id: I0289bee1f8e61fe61aedcf325e513196ff55de07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440062
Testability-Review: Brett Wilson <brettw@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>