[stash] Reduce severity of top level error logs
Reduces the severity of logs output when any error occurs while serving
a fidl client. These logs may be triggered if a client is shutdown
before stash during the teardown of a test. This can cause the test to
fail if the log is output with error severity.

Bug: 56611

Change-Id: Ia79c1071dfa7264dd2f3678d286a1258235a7351
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441482
Reviewed-by: Jody Sankey <jsankey@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Testability-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>