[fidl][rust] Populate next handle metadata on the encoder / decoder
This CL makes information on expected handle rights and subtype
available when it is time to encode or decode a handle. In
follow up CLs, the encoder and decoder will switch to use
HandleDisposition and HandleInfo and validate / encode rights.

Because there is no structure representing a "handle site",
this metadata needs to instead be generated on the struct,
tables and unions that contain handles. In a previous design,
this information would be passed through encode() calls until
a handle is reached, but for simplicity and likely performance
changes from the extra parameter the metadata is instead
attached to the Encoder and Decoder.

Change-Id: I1dc8d77c54a7f493196c8ff5e01ad15efc44b24b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/439600
Reviewed-by: Craig Tiller <ctiller@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>