[Build] Clean up update-to-zedboot.
This is part of the effort to remove zedboot from build that does
not use zedboot as recovery. According to etryzelaar@,
'update-to-zedboot' is no longer required so remove this.

I also don't see a usage of 'update.sources.manifest' so remove
that as well.

Tested: Presubmit.
Change-Id: I9f8817d5d0cd645e90b12a4f02ea68ece4d64b33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441721
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: Yu Shan <shanyu@google.com>