[ubsan] Disable vptr checks on host to unblock rollout
While the right thing to do here is to have up-to-date LLVM prebuilts
with variants, that is currently likely to take a while.  We can get
most of the upside of turning on UBSan on host tools by disabling this
particular check on host (where we do not have any UBSan automation
today) and following up.

Bug: 60446
Bug: 60253
Change-Id: If84f2a43b6e1438b167edac9a2d0b718e8e134ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441756
Commit-Queue: Drew Fisher <zarvox@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>