[hid-buttons] Refactor unit-tests
Refactor the unit test so that buttons are set up based on
their own configuration and not what type of test is being
run. This removes a lot of boilerplate but runs the test
in an identical way. It should make it easier to add
more tests in the future.

Bug: 62286
TEST: fx test hid-buttons-test

Change-Id: I67e33970e71792815428cec85f115fceff7396a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442061
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Testability-Review: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>