[web_runner_tests] Remove use_legacy_observer_behavior
It is an obsolete workaround to allow transition to the new behavior.

This is a manual revert of most of http://fxrev.dev/282890, which was
before the file was moved from topaz. The updated new behavior (when
use_legacy_observer_behavior is false) is retained.

Bug: 29937

Change-Id: Ia91d0dc86925dcddfff19c29bc5c80161bd69ea1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441853
Reviewed-by: Wez . <wez@google.com>
Testability-Review: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>