[kernel][vdso] Sync CPU startup during VDSO creation.
Make sure to wait until all CPUs have started on ARM before deciding
whether or not we need to install the A73 system timer read
mitigation.  Failure to do this would mean that we have a race, and we
might end up not installing the hook when we really need to do so.

Don't wait forever to make this decision.  If something goes terribly
wrong with CPU enumeration, simply install the mitigated read and take
the penalty whether we know that we need it or not.

Fixed: 59609
Change-Id: I8665c6b9cded8a4beeb6f2502e6b4b137ddeede3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/425313
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>