[build] Add `enable_variants` parameter to toolchain_variant_suite()
This adds a new parameter to toolchain_variant_suite() that allows
the caller to enable specific named variants, evern if they are not
listed in the global 'select_variant' global list.

This is necessary to build Asan-enabled variants of the C library
in the Fuchsia build, a requirement to move the Fuchsia sysroot
generation from the Zircon build.

+ Fix sanitizer_default_options.gni to properly handle Zircon-specific
  toolchains in the Fuchsia build.

Note that this only ensures that the GN toolchain() instances exist,
but does not add any dependencies to the build graph. For testing,
add the following to your universal_package_labels list:

  "//zircon/system/ulib/c(//zircon/system/ulib/c:user.libc_x64-asan)",
  "//zircon/system/ulib/c(//zircon/system/ulib/c:user.libc_x64-asan-ubsan)",

Then one can build the following targets with 'fx build':

  user.libc_x64-asan/libc.so
  user.libc_x64-asan-ubsan/libc.so

Fixed: 62555
Testing: See above.
Change-Id: I278ba52d3df62dd27e2d8b57f5782c3b72803501
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441821
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>