[fshost] Add options to fshost configs.
Previously the fshost build files were checked in and added as resources.
This change generates the build files when GN runs.

Adds a "default" set of options for the normal device target for the
urrent build. With the current settings this will contain only "default"
which will have no behavior impact. This changes the storage of the
options to no longer store "default" in the option set (this is handled
at a higher-level by the decoder).

Adds options for the max minfs and blobfs partition sizes. These are not
currently hooked up. Support is added for key/value options in the
fshost configuration.

Change-Id: Iafafcb19e60561e7414a86e0a1ae61346173bd25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441923
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>