[kernel][sched] Temporarily disable work stealing.
Temporarily disable work stealing in order to collect more data on
fxb/59720.

Right now, the main symptom that we are using in CQ in order to catch
repros of the stuck core bug is a situation where SSH stops working
because of a stuck thread, and the infrastructure reacts by running a
script over the UART to collect data for us.

Without work stealing in place, threads which get assigned to the
stuck core are stuck forever (or, at least for as long as the core is
stuck).  The scheduler is only going to stop sending threads to this
core when the estimated latency on the core is higher than all of the
other available options.  In high load situations, this can easily
cause 10+ threads to become stuck, and greatly increases the odds that
SSH will break as a result.

With work stealing in place, on the active thread on the stuck core ends up
stuck.  Other threads wait in the runnable queue until a core becomes
free, at which point in time the thread is stolen by the otherwise
idle core and run instead.  Clearly, this is not great behavior from a
thread scheduling latency point of view, but at least the thread
actually runs.

So, disable work stealing for the time being in the hope that it will
produce more really-stuck systems for us to collect data on.

Bug: 59720
Change-Id: If6e65fe8dfb165a9330fe419c9c54e4592790fa2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442414
Commit-Queue: Carlos Pizano <cpu@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>