[component_index] limit maximum number of results returned
On a build with a large number of packages (such as a build
from `fx set workstation.x64 --with=//bundles:tests`), the
number of components matching a query may exceed the
maximum size of a Zircon message. That causes the
component_index service to crash.

Fix this by setting a hard upper bound on the number of
matching items returned. Note that it's still possible for
the service to crash if the sum of the lengths of the
matching component names is large.

However, that problem is only hypothetical for now. So
we'll leave it for another day.

Bug: 45097
Test: fx test component_index_tests
Change-Id: I4244eab1f906b80e33e8b4770deb7ce85fa8f125
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442379
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Testability-Review: Shai Barack <shayba@google.com>
Testability-Review: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Mukesh Agrawal <quiche@google.com>