[fidl][sdk] Remove fidl-async from the SDK
//zircon/system/ulib/fidl-async contains a basic message dispatcher
implementation for the C bindings and LLCPP bindings. The library was
only included into the SDK because the HLCPP library declared it as a
dependency, which it doesn't need. Since we are deprecating
`fidl-async` and both C and LLCPP bindings are not intended to be used
from the SDK today, we should remove this library from the SDK by
removing the HLCPP dependency.

I did not find mentions of `fidl-async` / `fidl_async` in any partner
repos.

TEST: LSC presubmits

Change-Id: I19dc74ef0f1715583c0109e70e70612f5dbcbe72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442018
Reviewed-by: Pascal Perez <pascallouis@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Testability-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>