[wlan][policy] Exit when starting AP fails.
This change eliminates log spam in the event that starting
the AP fails.  Currently, a failure to start is
irrecoverable and the SME proxy is useless.  Exiting the
state machine allows the policy service and its clients the
opportunity to cleanup the interface and try again.

Bug: 61097
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: I047bc0dcec2190309ede125e479796d8bd744581
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433655
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Testability-Review: Marc Khouri <mnck@google.com>