[fuchsia-audio-codec] Fix off by one check on stream write
Fix check for space available for the case when space_left is equal to
the amount of bytes left in the current input buffer. This allows input
packets to be sent when exactly filled rather than waiting for the next
data to come in.

Change-Id: Ia9f09ed4d130ce7b687157c98676eeea7bc043f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/433873
Commit-Queue: Austin Foxley <afoxley@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>