[gn sdk] Remove bash -i from fserve.sh script to fix unexpected side effects
The #!/bin/bash -i was added in fxr/410994 and resulted in some
unexpected side effects. For example, femu.sh on OSX would suspend
itself on startup, somehow due to the -i creating a login shell that
changed the terminal properties. There were other problems with error
messages appearing in the output, although this was fixed in fxr/414742.
I suspect there were other bugs introduced by this as well.

However, the bash -i was originally introduced to fix a problem where
gsutils was not being found if in a non-system PATH location. Users will
need to add gsutils to a system PATH directory, and cannot rely on a custom
.bashrc or .bash_profile setting.

Change-Id: I96d6921f83b99a10010dc596860c46861e3c0ef2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440181
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Wayne Piekarski <waynepie@google.com>