[asan] Enable memory tests.
This CL re-enables the asan memory tests, which were previously disabled
due to the zero-page deduplication decomitting shadow memory in the
background.

The tests are now modified so that they are repeated if at least one
background decommit operation affected the asan shadow vmo.

The FillShadowPartialPages test was also refactored so we don't need to
run a for loop multiple times: now we just make sure that we are always
affecting partial shadow pages.

To build this test:

```
$ fx set bringup.x64 \
  --with-base //bundles/buildbot:bringup \
  --variant asan/sanitizer-utils \
  --with-base //zircon/system/ulib/c/test/sanitizer:sanitizer-utils
```

Run it with:

```
 runtests -r 1000 -n sanitizer-utils
```

Bug: 52653
Change-Id: I710f491fe538c4db7f32fa42ec93d8e2fc9011ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/436718
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>