[bt][a2dp-sink] Use Codec Negotiation
Use Codec Negotiation to select streams when we are initiating
streaming.  This should mean that we do not just select 44.1khz
which is not required to be supported by all sources.

It also means we select AAC when it is available in the source.

Fixed: 39321
Fixed: 48529
Test: added connect_initiation_uses_negotiation in ConnectedPeers
Change-Id: I9476082987b064345015bcc351b849e867659fa0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442154
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>