[bt][a2dp-source] Fix panic in encoding task
When the encoding AsyncWrite only had enough room for part of a buffer,
the encoding stream process would mark in the buffer where to resume,
push the buffer back on the queue, and then return to the top of the
loop to re-poll to register the waker.

The top of the loop would pop off the current buffer, and drop it on the
floor when it got a Poll::Pending.

This would also happen whenever the AsyncWrite for the encoder was
stalled, so we were also just occassionally dropping packets when the
encoder could take no packets.

Retain the buffer when we get Pending from the AsyncWrite.
Make the EncodedStream a bit more generic, and add a build_test which
doesn't require a StreamProcessor to operate the stream, making it
testable.

Fixed: fxbug.dev/62547
Test: added test_stalled_encoder_input in bt-a2dp-source-tests

Change-Id: I9214dc415ce533f30a9325ce6686bc2fd5eb25fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442996
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>