[hid-buttons] Refactor notifiers
Update the notifiers so that they rely on ButtonTypes
and are mapped by ButtonTypes. The notifiers are
listening to ButtonTypes via the FIDL API, so they
shouldn't be organized by BUTTON_IDS_

This has the added benefit of using the type system
more to make it clearer which ID's belong to what.

Bug: 36834
TEST: fx test hid-buttons-test

Change-Id: I7508c52fc7aa9eb500169f94b552232e59cf5183
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442199
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Testability-Review: Ruby Zhuang <rdzhuang@google.com>