[system-updater] Add a timeout to RebootController
As a precaution against misbehaving implementations of
fuchsia.update.Manager, this change adds an unreasonably long timeout on
waiting for the initiator of an update to signal that it is a good time
to reboot.

Also, this change reworks the RebootController to own the async task
waiting for the signal, which ties the lifecycle of the fidl protocol
handler for the reboot controller to the update attempt instead of the
client's connection, preventing misbehaving clients from keeping async
tasks around in the system-updater.

Alsoer, always provide a RebootController to the update attempt in the
fidl path, simplifying removing that Option<> later.

Fixed: 61956
Change-Id: I9c32d941f932f1ca0c321cdd63c9104d06fe6b6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442686
Testability-Review: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Zach Kirschenbaum <zkbaum@google.com>
Commit-Queue: Kevin Wells <kevinwells@google.com>