[wlan][policy] Expose more info from scan results
This CL's primary purpose is to route the information "did we observe
this network in a passive scan" from Network Selection through to the
Client State Machine, which (on successful connection) will record that
information in the Saved Networks Manager.

This information will help improve our determination of whether a
network is hidden or not.

The two main changes to enable this are:
1. within the Network Selection module, keep all the scanned bss
   information (rather than simplifying to only a single RSSI).
2. include a "network metadata" parameter in the ConnectRequest,
   which contains the "network observed in a passive scan" boolean.

Bug: 35918
Change-Id: Ic5e70237c5245117298e0f9c06578cfd45e0ca97
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441855
Commit-Queue: Marc Khouri <mnck@google.com>
Reviewed-by: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Testability-Review: Kevin Sakuma <sakuma@google.com>