[fvm] Add oldest_revision to FVM header.
This implements the policy in //src/storage/docs/versioning.md for
tracking the oldest revision of the software that has touched the data.
The current revision is set to 1 so we can differentiate headers
produced before this change. The format is unchanged and is
backwards-compatible with old versions of the driver.

The existing "version" field is renamed "format_version" to align with
the document's terminology but its meaning and binary representation is
unchanged.

Update FVM driver test to write a real FVM header to the udnerlying fake
block device and initialize the FVM driver like it would be in normal
operation. This allows removing some test only code, allows testing of
Load(), and makes the tests generally more realistic.

Adds tests for oldest_revision updates.

Bug: 62594

Change-Id: I3a4aefb8aaee9133ac0255e0c990fe81d22d63a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442564
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>