[blobfs][chunked-compress] Add Simplified DecompressFrame.
Adds a method to chunked-compression for doing a simplified call here
the caller is responsible for ensuring that the start points and lengths
encapsulate a complete frame rather than using the seek table to
describe them.

This is used in the remote decompressor where the remote process has no
information about the seek tables and performs the decompression relying
on blobfs to give it the correct values.

BUG: 62396
Change-Id: Ie5389443d8dcf21245cfee84d6020e370586c9fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/442628
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>