[pkg-resolver-integration-tests] Fix flake w/ inspect race condition
Originally, we were just waiting for the unblocker to be sent by the
http server, which guarantees the pkg-resolver has made the GET
request for the blob, but not that the pkg-resolver is waiting for
HTTP body bytes (it could be e.g. still waiting for header bytes or
truncating the blob in blobfs).

MULTIPLY: fuchsia-pkg://fuchsia.com/pkg-resolver-integration-tests#meta/pkg-resolve-inspect.cmx
Fixed: 60578
Change-Id: I1a017e4908b54eb4c56adacc001a33971579e9e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443265
Reviewed-by: John Wittrock <wittrock@google.com>
Testability-Review: Ben Keller <galbanum@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>