[kernel][vm] Introduce VM hierarchy base type
This moves the references to the VMO hierarchy state as well as methods
for retriving the locks to a base type. The only actual effect this has
is that the memory layout of the VmObject struct is possibly different,
but otherwise there are no semantic changes.

Motivation for this is to prepare the way for additional objects to
be part of the VMO hierarchy and share the lock and other state,
without having to be based on VmObject.

Change-Id: I9b8c9a761cfdb45548a81538e18fc0eb8e2121e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437340
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>