[kernel][vm] Introduce VmCowPages
This creates a parallel hierarchy to VmObjectPaged in VmCowPages. The
logic of actually manipulating copy-on-write through the VmPageList is
handled by VmCowPages, with VmObjectPaged maintaining responsibility
for handling VmMappings and providing the user API.

Standalone this commit is a temporary regression in terms of memory use
and performance due to the VmObjectPaged still creating needless hidden
parents that serve no purpose. The reason to have a temporary regression
is to keep this commit as close to "moving code" as possible without
there also being additional complex changes hidden inside. This does
lay the groundwork for successive commits to perform the 'interesting'
transformations to end up in an eventually more efficient hierarchy.

Change-Id: I32ee7e8f1a158aa3e8cfa3cdb87435ec129588e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437341
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>