[kernel][vm] Hold lock over VmCowPages::Commit
This redistributes the commit logic between VmCowPages and
VmObjectPaged. The goal is to move the top level retry loop that has to
wait on page requests with the lock dropped in VmObjectPaged, and have
the VmCowPages perform only the section that can be done with the lock
held.

Motivation for this is that later we would like the cow_pages_ link
in VmObjectPaged to be modifiable to point to different VmCowPages. If
VmCowPages would be dropping the lock mid operation then it could
potentially cease being the linked object, and  when the lock is
reacquired be acting on the wrong VmCowPages. Dropping the lock in
VmObjectPaged ensures that after we re-acquire we will use any
potentially updated cow_pages_ link.

Change-Id: I8ab81aa621511013b2846b47fa460fd6d7be1048
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437345
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>