[kernel][vm] Lock access to VmObjectPaged::cow_pages_
This changes cow_pages_ to no longer be const and require the lock
being held to reference it. As a consequence there are additional
VmCowPages operations need to be externally locked as VmObjectPaged
needs to hold the lock prior to being able to call the operation.

The motivation for holding the lock to access cow_pages_ is so that
later we can modify it over the lifetime of a VmObjectPaged. This
creates complexity for operations such as CreateClone and
CreateChildSlice that currently rely on performing some actions outside
of the lock to prevent deadlock.

To hold the lock over these operations we allow cow_pages_ to be null
on VmObjectPaged creation. This allows us to create a VmObjectPaged
optimistically, then within one lock acquisition create the VmCowPages
child and setup all the backlinks.

cow_pages_ being null becomes a special condition that we do not want
to have to generally deal with and so we declare that cow_pages_ should
only be null during VmObjectPaged creation and that all methods are able
to assume it is not null. Currently a VMO is usable prior to the
creation request completing by virtue of all VMOs being added to a
global list. To prevent VMOs being found in the global list prior to
cow_pageS_ being set the insertion into the list is externalized,
instead of performing it in the constructor.

Change-Id: I00f65962ddf6aee9f3f8be7cf6333720a1bffda3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437346
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>