[kernel][vm] Remove hidden parents from VmObjectPaged
This removes the notion of hidden parents from VmObjectPaged and in
the process places control of the VmCowPages hierarchy manipulation in
the VmCowPages code, instead of being done from the VmObjectPaged.

Aside from the removal of the redundant VmObjectPaged hidden parents
this also performs an additional optimization of making the parent_
reference in VmObjectPaged a raw reference and not a RefPtr. This
can be done since now that the copy-on-write pages are held in
VmCowPages there is no reason to hold the VmObjectPaged around after
any other references, such as user handles and mappings, have gone
away. Not holding the parent via a refptr also ensures we can not have
any chained deletions and so the deferred deletion mechanism is only
needed now by VmCowPages.

Change-Id: I710eab49b583542f2a773f130319d5b4a8dfc7e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/437347
Testability-Review: Adrian Danis <adanis@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>