[jitterentropy] Break dependency on arch/x86/feature.h for InvTSC bit
Jitterentropy is a third party C library linked into the kernel to serve
as an entropy source for the kernel csprng. Unlike much of the rest of
the kernel, it is compiled as C source.

On x86 systems, it tests for an Invariant TSC, to decide how to
incorporate timing information. Convert its use of x86_feature_test to a
raw call to CPUID - this allows us to remove the inclusion of zircon
header files, which allows us to make those headers C++-only.

Bug: 47117 Use only ktl::atomic in the kernel

Change-Id: I1570b83934531ea77704e547e2db6ef90afb96b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/441822
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>