[SDK] Converting fpublish.sh to binary executable
This creates an sdk-tools project under //tools

The tool fpublish.sh is implemented as a bash shell and does not
meet the CLI rubric requirement with respect to implmentation language.

This ports the bash script to Go.

This is a temporary solution until ffx package management is completed.

Change-Id: I949a98360353552c957682721764633283f555f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/434917
Reviewed-by: Sean Cuff <seancuff@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Testability-Review: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>