[check-licenses] Fixing license propagation
With fxr/443878, I can see an accurate picture of the fuchsia tree.
The part of the code that collects LICENSE files and the part of the
code that verifies the contents of the file were out of sync: we
shouldn't count files as licenses if their contents can't be
verified. This is now fixed in traverse.go, lines 146-162.

Also, we shouldn't assume the root Fuchsia LICENSE file applies to
projects under prebuilt / third_party directories. I added support
for only propagating licenses up to this directory name boundary.
The boundaries are configurable in config.json, under stopLicensePropagation.

Bug: b/171843674

Change-Id: I57524ff537bd2942c7fe1d97accc28ffc50a107d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443939
Testability-Review: Francois Rousseau <frousseau@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>