[bt][fidl][gatt] Fix gatt.Client lifetimes
GattHost used to maintain a 1:1 mapping gatt.Client handles and a
single owning le.Central instance. This meant that:

1. A given le.Central capability could only maintain a GATT FIDL
connection to one peer at a time.
2. A peer disconnection (either remotely or locally initiated) caused
the tear down of all other peers attached to any associated le.Central
capability.

This was due to incorrect semantics in GattHost's internal logic as well
as its public methods. This is now fixed, such that a gatt.Client now
maps to an owning le.Central AND the ID of the peer it's associated
with.

A subsequent fix should remove this GattHost indirection altogether,
as detailed in fxbug.dev/62940.

Bug: 62941
Test: bt-host-unittests
Change-Id: I7810a7e8c1ab8e7ad9cbb2c5af23d15a673cdedf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444694
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Testability-Review: Ben Lawson <benlawson@google.com>