[kernel][x86] Move L1D$ flush to before thread-signaled check
When returning from a syscall with an error result we flush the L1D$
to mitigate Spectre V1.

Prior to this change, we performed the error result check and flush
after a debugger had a chance to modify the returning thread's
registers.  This meant there was an opportunity for a debugger to
change an error result to a non-error result and therefore bypass the
L1D$ flush.

This change moves the check and flush earlier in the syscall return
path to before the thread signals are checked.

Change-Id: I90984e326605bf89aeb3c124d22c597f6260724f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/443315
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>