[bt][rfcomm] Handle DisconnectedMode responses
Adds support to handle DM responses from the remote peer. A DM response
is typically received as a form of rejection or error. For a user DLC, a
DM response typically means the DLC is in a bad state and should be
closed. For the mux control DLC, a DM response is typically sent if
the remote peer rejects multiplexer startup.

There is also a special case in which the peer can respond to a
Parameter Negotiation Mux command with a DM or DISC  over the DLC
referred to in the PN command. This special case is defined in RFCOMM
5.5.3. Adds support for this special case + tests.

Bug: 59585
Test: Added to bt-rfcomm-tests
Change-Id: I823d8f1b5e9983c70fee0ce85a425405c40128fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/444215
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>
Testability-Review: Lucas Jenkins <lucasjenkins@google.com>